pyproject.toml
[project]
name = "3-python-vibe-coding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
]
